+++ src/main/java/act/job/JobExceptionListenerFinder.java
@@ -1,0 +1,31 @@
+package act.job;
+
+/*-
+ * #%L
+ * ACT Framework
+ * %%
+ * Copyright (C) 2014 - 2019 ActFramework
+ * %%
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * #L%
+ */
+
+import act.Act;
+import act.util.SubClassFinder;
+
+public class JobExceptionListenerFinder {
+    @SubClassFinder
+    public void found(JobExceptionListener l) {
+        Act.jobManager().exceptionListenerManager().registerJobExceptionListener(l);
+    }
+}